wikiversionchange

+++ WIKI/includes/db/CloneDatabase.php
@@ -2,6 +2,9 @@
 /**
  * Helper class for making a copy of the database, mostly for unit testing.
  *
+ * Copyright © 2010 Chad Horohoe <chad@anyonecanedit.org>
+ * https://www.mediawiki.org/
+ *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation; either version 2 of the License, or
@@ -21,21 +24,20 @@
  * @ingroup Database
  */
 use MediaWiki\MediaWikiServices;
-use Wikimedia\Rdbms\ILoadBalancer;
 use Wikimedia\Rdbms\IMaintainableDatabase;
 
 class CloneDatabase {
 	/** @var string Table prefix for cloning */
-	private $newTablePrefix;
+	private $newTablePrefix = '';
 
 	/** @var string Current table prefix */
-	private $oldTablePrefix;
+	private $oldTablePrefix = '';
 
 	/** @var array List of tables to be cloned */
-	private $tablesToClone;
+	private $tablesToClone = [];
 
 	/** @var bool Should we DROP tables containing the new names? */
-	private $dropCurrentTables;
+	private $dropCurrentTables = true;
 
 	/** @var bool Whether to use temporary tables or not */
 	private $useTemporaryTables = true;
@@ -47,19 +49,16 @@ class CloneDatabase {
 	 * @param IMaintainableDatabase $db A database subclass
 	 * @param array $tablesToClone An array of tables to clone, unprefixed
 	 * @param string $newTablePrefix Prefix to assign to the tables
-	 * @param string|null $oldTablePrefix Prefix on current tables, if not $wgDBprefix
+	 * @param string $oldTablePrefix Prefix on current tables, if not $wgDBprefix
 	 * @param bool $dropCurrentTables
 	 */
 	public function __construct( IMaintainableDatabase $db, array $tablesToClone,
-		$newTablePrefix, $oldTablePrefix = null, $dropCurrentTables = true
+		$newTablePrefix, $oldTablePrefix = '', $dropCurrentTables = true
 	) {
-		if ( !$tablesToClone ) {
-			throw new InvalidArgumentException( 'Empty list of tables to clone' );
-		}
 		$this->db = $db;
 		$this->tablesToClone = $tablesToClone;
 		$this->newTablePrefix = $newTablePrefix;
-		$this->oldTablePrefix = $oldTablePrefix ?? $this->db->tablePrefix();
+		$this->oldTablePrefix = $oldTablePrefix ? $oldTablePrefix : $this->db->tablePrefix();
 		$this->dropCurrentTables = $dropCurrentTables;
 	}
 
@@ -86,29 +85,31 @@ class CloneDatabase {
 			# works correctly across DB engines, we need to change the pre-
 			# fix back and forth so tableName() works right.
 
-			$this->db->tablePrefix( $this->oldTablePrefix );
+			self::changePrefix( $this->oldTablePrefix );
 			$oldTableName = $this->db->tableName( $tbl, 'raw' );
 
-			$this->db->tablePrefix( $this->newTablePrefix );
+			self::changePrefix( $this->newTablePrefix );
 			$newTableName = $this->db->tableName( $tbl, 'raw' );
 
 			// Postgres: Temp tables are automatically deleted upon end of session
 			//           Same Temp table name hides existing table for current session
-			if ( $this->dropCurrentTables ) {
+			if ( $this->dropCurrentTables
+				&& !in_array( $this->db->getType(), [ 'oracle' ] )
+			) {
 				if ( $oldTableName === $newTableName ) {
 					// Last ditch check to avoid data loss
 					throw new LogicException( "Not dropping new table, as '$newTableName'"
 						. " is name of both the old and the new table." );
 				}
 				$this->db->dropTable( $tbl, __METHOD__ );
-				wfDebug( __METHOD__ . " dropping {$newTableName}" );
+				wfDebug( __METHOD__ . " dropping {$newTableName}\n" );
 				// Dropping the oldTable because the prefix was changed
 			}
 
 			# Create new table
-			wfDebug( __METHOD__ . " duplicating $oldTableName to $newTableName" );
+			wfDebug( __METHOD__ . " duplicating $oldTableName to $newTableName\n" );
 			$this->db->duplicateTableStructure(
-				$oldTableName, $newTableName, $this->useTemporaryTables, __METHOD__ );
+				$oldTableName, $newTableName, $this->useTemporaryTables );
 		}
 	}
 
@@ -118,34 +119,25 @@ class CloneDatabase {
 	 */
 	public function destroy( $dropTables = false ) {
 		if ( $dropTables ) {
-			$this->db->tablePrefix( $this->newTablePrefix );
+			self::changePrefix( $this->newTablePrefix );
 			foreach ( $this->tablesToClone as $tbl ) {
-				$this->db->dropTable( $tbl, __METHOD__ );
+				$this->db->dropTable( $tbl );
 			}
 		}
-		$this->db->tablePrefix( $this->oldTablePrefix );
+		self::changePrefix( $this->oldTablePrefix );
 	}
 
 	/**
-	 * Change the table prefix on all open DB connections
+	 * Change the table prefix on all open DB connections/
 	 *
 	 * @param string $prefix
 	 * @return void
 	 */
 	public static function changePrefix( $prefix ) {
-		global $wgDBprefix, $wgDBname;
+		global $wgDBprefix;
 
 		$lbFactory = MediaWikiServices::getInstance()->getDBLoadBalancerFactory();
-		$lbFactory->setLocalDomainPrefix( $prefix );
-
-		$aliases = [
-			$wgDBname => $lbFactory->getLocalDomainID()
-		];
-		$lbFactory->setDomainAliases( $aliases );
-		$lbFactory->forEachLB( function ( ILoadBalancer $lb ) use ( $aliases ) {
-			$lb->setDomainAliases( $aliases );
-		} );
-
+		$lbFactory->setDomainPrefix( $prefix );
 		$wgDBprefix = $prefix;
 	}
 }
